requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain-community
python-dotenv